Refactor repack to AbstractArticleCrudService
